back/requirements.txt
python-arango == 7.5.7
Flask == 2.0.3
flask_cors == 3.0.8
connexion == 2.6.0
requests == 2.25.1
swagger-ui-bundle == 0.0.5
openapi_spec_validator == 0.2.8
PyYAML == 6.0
termcolor==2.3.0